added category service tests

+++ src/main/java/com/example/bookstore/service/BookServiceImpl.java
@@ -49,6 +49,9 @@ public class BookServiceImpl implements BookService {
 
     @Override
     public void deleteById(Long id) {
+        if (!bookRepository.existsById(id)) {
+            throw new EntityNotFoundException("Can't delete a book by id " + id);
+        }
         bookRepository.deleteById(id);
     }
